refactor : MVC 패턴으로 파일 분리
