+++ src/App.js
@@ -1,14 +1,15 @@
-import { MissionUtils } from "@woowacourse/mission-utils";
-import LottoPurchase from "./LottoPurchase";
-import LottoTicket from "./LottoTicket";
-import LottoMachine from "./LottoMachine";
+import { MissionUtils } from '@woowacourse/mission-utils';
+import LottoPurchase from './Model/LottoPurchase';
+import LottoTicket from './Model/LottoTicket';
+import LottoMachine from './Model/LottoMachine';
+import LottoOutput from './View/LottoOutput';
 
 class App {
   async play() {
     const PURCHASE_AMOUNT = await LottoPurchase.purchaseLottoAmount();
     const LOTTO_TICKETS = LottoTicket.generateLottotickets(PURCHASE_AMOUNT);
 
-    LottoMachine.displayLottoTickets(LOTTO_TICKETS);
+    LottoOutput.displayLottoTickets(LOTTO_TICKETS);
 
     const WINNING_NUMBERS = await LottoMachine.askWinningNumbers();
     const BONUS_NUMBER = await LottoMachine.askBonusNumber();
@@ -16,10 +17,10 @@ class App {
     const PRIZE_LIST = LottoMachine.calculatePrize(
       LOTTO_TICKETS,
       WINNING_NUMBERS,
-      BONUS_NUMBER
+      BONUS_NUMBER,
     );
 
-    LottoMachine.displayGameResult(PRIZE_LIST, PURCHASE_AMOUNT);
+    LottoOutput.displayGameResult(PRIZE_LIST, PURCHASE_AMOUNT);
   }
 }
