--- src/LottoOutput.js
@@ -1,3 +1,0 @@
-import { MissionUtils } from '@woowacourse/mission-utils';
-
-class LottoOutput {}
